Triggers -> Edit from Detail View
Added a dialog that allows you to edit your trigger from the detail
view. This is a test and has some quirks. This was to prove that it can
be done so that we can update the edit functionality a bit to make it
easier on the user (no more go to this page, go to that page, etc).
Things that don't work with it:
- You can cause multiple dialogs to pop up
- You cannot see the update in the detail view after the update is
successful
- There's no update to the UI to indicate it was successful from the
form dialog

Untested:
- Has not been tested on mobile devices (phone or tablet or surface pro
2)